refactor: unify work-tile candidacy and declarative standard-work config (#3877)
Consolidate prefilter and tile-keys logic into order_suggestion_work_tile_candidacy.dart
with thin re-export shims; table-drive build_improvement and build_fort config builders.
