+++ packages/colonizethis_orders/lib/src/orders/order_suggestion_work_worker.dart
@@ -6,8 +6,7 @@ import 'feedstock_extraction_targets.dart'
     show feedstockExtractionResourceIdsForPlayer;
 import 'incremental_candidate_validator.dart';
 import 'order_suggestion_context.dart';
-import 'order_suggestion_work_tile_keys.dart';
-import 'order_suggestion_work_tile_prefilter.dart';
+import 'order_suggestion_work_tile_candidacy.dart';
 import 'order_work_constants.dart';
 import 'unit_type_helpers.dart';
 import 'work_suggestion_pipeline.dart';
